Add files via upload
This GitHub repository provides an analysis of theft crime data for the years 2020 to 2023 across different neighborhoods. The top three neighborhoods with the least amount of theft crime are Foothill, Harbor, and Hollenbeck, with decreasing crime rates over the specified years. In contrast, Pacific, West LA, and Central consistently have higher numbers of theft crimes. While crime rates are considered, it's emphasized that they alone may not provide a comprehensive overview of neighborhood livability. The repository aims to offer valuable insights for individuals seeking information on crime trends in specific neighborhoods, with a reminder to consider other factors like community amenities and cost of living in the decision-making process.